Make --no-download install if all files present
Previously, no-download functioned like a dry-run, it would never
actually install. This was a bit problematic as it meant there were
still no "truly offline" installation path.

This change makes it so `--no-download` will ask the user to store the
files if they're not all present, and will actually run the download if
the files are present.

The nice thing is that this codepath is also active outside no-download,
so if multiple files fail to download, there will be a nice little
summary of all the files that failed.